fixes for Courierdashboard
