+++ frontend/src/layout/Sidebar.jsx
@@ -2,72 +2,154 @@ import React, { useEffect, useState } from "react";
 import { NavLink } from "react-router-dom";
 import { jwtDecode } from "jwt-decode";
 
-export default function Sidebar() {
+export default function Sidebar({
+  collapsed,
+  setCollapsed,
+  navbarHeight = 64,
+  expandedWidth = 260,
+  collapsedWidth = 88,
+  sidebarRef,
+}) {
   const [role, setRole] = useState("USER");
 
   useEffect(() => {
     const token = localStorage.getItem("accessToken");
     if (token) {
       try {
-        const decoded = jwtDecode(token);
-        const userRole = decoded.roles?.[0] || "USER";
-        setRole(userRole);
-      } catch (err) {
-        console.error("Invalid token:", err);
-      }
+        const d = jwtDecode(token);
+        setRole((d.roles?.[0] || "USER").toUpperCase());
+      } catch {}
     }
   }, []);
 
+  /** ---------------- Role → Nav items map ---------------- */
+  const baseItems = [
+    { path: "/dashboard", label: "Dashboard", icon: "bi-grid" },
+  ];
+
+  const roleNavMap = {
+    COURIER: [
+      { path: "/orders", label: "Previous Orders", icon: "bi-clock-history" },
+      { path: "/profile", label: "Profile", icon: "bi-person" },
+    ],
+    RECEIVER: [
+      { path: "/profile", label: "Profile", icon: "bi-person" },
+      { path: "/requests", label: "My Requests", icon: "bi-basket" },
+      { path: "/delivery", label: "Delivery Update", icon: "bi-truck" },
+      { path: "/orders", label: "Previous Orders", icon: "bi-clock-history" },
+    ],
+    DONOR: [
+      { path: "/profile", label: "Profile", icon: "bi-person" },
+      { path: "/lots", label: "My Lots", icon: "bi-box-seam" },
+      { path: "/bulk-import", label: "Bulk Import", icon: "bi-cloud-upload" },
+    ],
+    DEFAULT: [
+      { path: "/profile", label: "Profile", icon: "bi-person" },
+    ],
+  };
+
   const items = [
-    { path: "/dashboard", label: "Dashboard", icon: "bi-speedometer2" },
-    role === "DONOR" && {
-      path: "/donations",
-      label: "My Donations",
-      icon: "bi-box-seam",
-    },
-    role === "RECEIVER" && {
-      path: "/requests",
-      label: "My Requests",
-      icon: "bi-basket",
-    },
-    role === "COURIER" && {
-      path: "/jobs",
-      label: "Delivery Jobs",
-      icon: "bi-truck",
-    },
-    { path: "/profile", label: "Profile", icon: "bi-person" },
-  ].filter(Boolean);
+    ...baseItems,
+    ...(roleNavMap[role] ?? roleNavMap.DEFAULT),
+  ];
+
+  const handleLogout = async () => {
+    const accessToken = localStorage.getItem("accessToken");
+    const refreshToken = localStorage.getItem("refreshToken");
+    try {
+      await fetch("http://localhost:8080/api/v1/auth/logout", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${accessToken}`,
+          "X-Refresh-Token": refreshToken || "",
+        },
+      });
+    } catch {}
+    // localStorage.removeItem("accessToken");
+    // localStorage.removeItem("refreshToken");
+    // window.location.href = "/authentication";
+  };
+
+  // Expand sidebar on any click inside it while collapsed.
+  const handleRootMouseDown = (e) => {
+    if (!collapsed) return;
+    if (e.target.closest('[data-action="logout"]')) return;
+    setCollapsed(false);
+    if (e.target.closest("a, button")) e.preventDefault();
+    e.stopPropagation();
+  };
 
   return (
     <aside
-      className="d-flex flex-column bg-secondary bg-opacity-10 border-end shadow-sm"
-      style={{ width: 240 }}
+      ref={sidebarRef}
+      onMouseDown={handleRootMouseDown}
+      aria-expanded={!collapsed}
+      title={collapsed ? "Expand sidebar" : undefined}
+      style={{
+        position: "fixed",
+        top: "var(--navbar-h)",
+        left: 0,
+        height: `calc(100vh - var(--navbar-h))`,
+        width: collapsed ? `${collapsedWidth}px` : `${expandedWidth}px`,
+        backgroundColor: "#fff",
+        borderRight: "1px solid rgba(0,0,0,.1)",
+        transition: "width 0.25s ease",
+        zIndex: 1050,
+        cursor: collapsed ? "pointer" : "default",
+        userSelect: "none",
+      }}
+      className="d-flex flex-column"
     >
-      <div className="p-3 border-bottom">
-        <h6 className="text-uppercase text-muted mb-0">
-          {role === "DONOR" && "Donor Portal"}
-          {role === "RECEIVER" && "Receiver Portal"}
-          {role === "COURIER" && "Courier Portal"}
-          {role === "USER" && "User Portal"}
-        </h6>
-      </div>
-
-      <nav className="flex-grow-1 mt-2">
+      <nav className="flex-grow-1 py-3">
         {items.map((it, i) => (
           <NavLink
             key={i}
             to={it.path}
+            end
+            tabIndex={collapsed ? -1 : 0}
             className={({ isActive }) =>
-              `d-flex align-items-center px-3 py-2 text-decoration-none ${
-                isActive ? "bg-primary text-white rounded" : "text-light"
+              `d-flex align-items-center text-decoration-none mb-2 ${
+                isActive ? "text-white" : "text-secondary"
               }`
             }
+            style={({ isActive }) => ({
+              height: 48,
+              borderRadius: 10,
+              transition: "all 0.25s ease",
+              padding: collapsed ? "0 12px" : "0 18px",
+              margin: "0 12px",
+              justifyContent: collapsed ? "center" : "flex-start",
+              fontWeight: isActive ? 600 : 500,
+              backgroundColor: isActive ? "#111827" : "transparent",
+              pointerEvents: collapsed ? "none" : "auto",
+            })}
           >
-            <i className={`bi ${it.icon} me-2`} />
-            {it.label}
+            <i
+              className={`bi ${it.icon}`}
+              style={{ fontSize: "1.2rem", marginRight: collapsed ? 0 : 12 }}
+            />
+            {!collapsed && (
+              <span className="fw-medium" style={{ fontSize: "0.95rem" }}>
+                {it.label}
+              </span>
+            )}
           </NavLink>
         ))}
       </nav>
+
+      <div className="border-top p-2 d-flex justify-content-center">
+        <button
+          className="btn btn-link text-danger d-flex align-items-center gap-2 px-2"
+          data-action="logout"
+          onClick={handleLogout}
+          style={{ textDecoration: "none", fontWeight: 600 }}
+          title="Logout"
+        >
+          <i className="bi bi-box-arrow-right" />
+          {!collapsed && <span>Logout</span>}
+        </button>
+      </div>
     </aside>
   );
 }
