+++ frontend/src/routes/AppRoutes.jsx
@@ -1,7 +1,6 @@
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import SignupPage from "../pages/signup/SignupPage";
 import CourierVerificationPage from "../pages/CourierVerificationPage";
-import CourierDashboard from "../pages/CourierDashboard";
 
 import UserDashboard from "../pages/dashboards/userDashboard/UserDashboard";
 
@@ -12,7 +11,6 @@ function AppRoutes() {
         <Route path="/" element={<Navigate to="/authentication" replace />} />
         <Route path="/authentication" element={<SignupPage />} />
         <Route path="/courier-verification" element={<CourierVerificationPage />} />
-        <Route path="/courier-dashboard" element={<CourierDashboard />} />
         {/* <Route path="/jobs" element={<JobsPage />} /> */}
         <Route path="/dashboard" element={<UserDashboard />} />
       </Routes>
